add PosRec to event_basics and move to different module
