+++ straxen/plugins/peak_processing.py
@@ -1,8 +1,6 @@
 import json
 import os
 import tempfile
-import tarfile
-
 import numpy as np
 import numba
 
@@ -113,148 +111,6 @@ class PeakBasicsHighEnergy(PeakBasics):
 
 @export
 @strax.takes_config(
-    strax.Option('min_reconstruction_area',
-                 help='Skip reconstruction if area (PE) is less than this',
-                 default=10),
-    strax.Option('n_top_pmts', default=straxen.n_top_pmts,
-                 help="Number of top PMTs")
-)
-class PeakPositionsBaseNT(strax.Plugin):
-    """
-    Base class for reconstructions.
-    This class should only be used when
-    """
-    depends_on = ('peaks',)
-    algorithm = None
-    # TODO
-    #  check that this works properly, significant loading/saving penalty for three plugins!
-    parallel = 'process'
-    __version__ = '0.0.0'
-
-    def infer_dtype(self):
-        if self.algorithm is None:
-            raise NotImplementedError(f'Base class should not be used without '
-                                      f'algorithm as done in {__class__}')
-        dtype = [('x_'+self.algorithm, np.float32,
-                  f'Reconstructed {self.algorithm} S2 X position (cm), uncorrected'),
-                 ('y_'+self.algorithm, np.float32,
-                  f'Reconstructed {self.algorithm} S2 Y position (cm), uncorrected')]
-        dtype += strax.time_fields
-        return dtype
-
-    def setup(self):
-        if self.config["file_"+self.algorithm] is None:
-            raise ValueError(f'No file provided to reconstruction with for {self.algorithm}')
-        import tensorflow as tf
-        self.model_file = str(self.config["file_"+self.algorithm])
-        self.model = None
-        if os.path.exists(self.model_file):
-            print(f"Path is local. Loading {self.algorithm} TF model locally from disk.")
-        else:
-            downloader = straxen.MongoDownloader()
-            if self.model_file in downloader.list_files():
-                self.model_file = downloader.download_single(self.model_file)
-            else:
-                raise RuntimeError("Model files {0} is not found".format(self.model_file))
-        with tempfile.TemporaryDirectory() as tmpdirname:
-            tar = tarfile.open(self.model_file, mode="r:gz")
-            tar.extractall(path=tmpdirname)
-            self.model = tf.keras.models.load_model(tmpdirname)
-
-    def compute(self, peaks):
-        result = np.ones(len(peaks), dtype=self.dtype)
-        result['time'], result['endtime'] = peaks['time'], strax.endtime(peaks)
-
-        result['x_'+self.algorithm] *= float('nan')
-        result['y_'+self.algorithm] *= float('nan')
-        # Keep large peaks only
-        peak_mask = peaks['area'] > self.config['min_reconstruction_area']
-        if not np.sum(peak_mask):
-            # Nothing to do, and .predict crashes on empty arrays
-            return result
-        
-        # Getting actual position reconstruction
-        _in = peaks['area_per_channel'][peak_mask, 0:self.config['n_top_pmts']]
-        with np.errstate(divide='ignore', invalid='ignore'):
-            _in = _in / np.max(_in, axis=1).reshape(-1, 1)
-        _in = _in.reshape(-1, self.config['n_top_pmts'])
-        _out = self.model.predict(_in) 
-                
-        # writing output to the result 
-        result['x_'+self.algorithm][peak_mask] = _out[:, 0]
-        result['y_'+self.algorithm][peak_mask] = _out[:, 1]              
-        return result
-
-
-@export 
-@strax.takes_config(
-    strax.Option("file_mlp", 
-                 help="Name of saved MLP model file in the aux file data base."
-                      "The file contains both structure and weights.",
-                 default=None
-                 )
-)
-class PeakPositionsMLP(PeakPositionsBaseNT):
-    provides = "peak_positions_mlp"
-    algorithm = "mlp"
-
-
-@export 
-@strax.takes_config(
-    strax.Option("file_gcn", 
-                 help="Name of saved GCN model file in the aux file data base."
-                      "The file contains both structure and weights.",
-                 default=None
-                 )
-)
-class PeakPositionsGCN(PeakPositionsBaseNT):
-    provides = "peak_positions_gcn"
-    algorithm = "gcn"
-
-
-@export 
-@strax.takes_config(
-    strax.Option("file_cnn", 
-                 help="Name of saved CNN model file in the aux file data base."
-                      "The file contains both structure and weights.",
-                 default=None
-                 )
-)
-class PeakPositionsCNN(PeakPositionsBaseNT):
-    provides = "peak_positions_cnn"
-    algorithm = "cnn"
-
-
-@export
-@strax.takes_config(
-    strax.Option("default_reconstruction_algorithm",
-                 help="default reconstruction algorithm that provides (x,y)",
-                 default="mlp",
-                 )
-)
-class PeakPositionsNT(strax.MergeOnlyPlugin):
-    provides = "peak_positions"
-    depends_on = ("peak_positions_cnn", "peak_positions_mlp", "peak_positions_gcn")
-    save_when = strax.SaveWhen.NEVER
-
-    def infer_dtype(self):
-        dtype = strax.merged_dtype([self.deps[d].dtype_for(d) for d in self.depends_on])
-        dtype += [('x', np.float32, 'Reconstructed S2 X position (cm), uncorrected'),
-                  ('y', np.float32, 'Reconstructed S2 Y position (cm), uncorrected')]
-        return dtype
-
-    def compute(self, peaks):
-        result = {dtype: peaks[dtype] for dtype in peaks.dtype.names}
-        algorithm = self.config['default_reconstruction_algorithm']
-        if not 'x_' + algorithm in peaks.dtype.names:
-            raise ValueError
-        for xy in ('x', 'y'):
-            result[xy] = peaks[f'{xy}_{algorithm}']
-        return result
-
-
-@export
-@strax.takes_config(
     strax.Option(
         'nn_architecture',
         help='Path to JSON of neural net architecture',
@@ -267,11 +123,6 @@ class PeakPositionsNT(strax.MergeOnlyPlugin):
         default_by_run=[
             (0, pax_file('XENON1T_tensorflow_nn_pos_weights_20171217_sr0.h5')),
             (first_sr1_run, pax_file('XENON1T_tensorflow_nn_pos_weights_20171217_sr1.h5'))]),   # noqa
-    strax.Option("nn_file",
-                 help="Name of saved neural network model file in the aux file "
-                      "data base. The file contains both structure and weights.",
-                 default=None
-                    ),
     strax.Option('min_reconstruction_area',
                  help='Skip reconstruction if area (PE) is less than this',
                  default=10),
@@ -301,22 +152,6 @@ class PeakPositions1T(strax.Plugin):
     def setup(self):
         import tensorflow as tf
         keras = tf.keras
-
-        # XENONnT
-        if self.config['nn_file']:
-            downloader = straxen.MongoDownloader()
-            tf.keras.backend.set_floatx('float32')
-
-            nn_file = downloader.download_single(self.config['nn_file'])
-
-            with tempfile.TemporaryDirectory() as tmpdirname:
-                tar = tarfile.open(nn_file, mode="r:gz")
-                tar.extractall(path=tmpdirname)
-                self.nn = tf.keras.models.load_model(tmpdirname)
-            # Done, let's return
-            return
-
-        # XENON1T
         nn_conf = get_resource(self.config['nn_architecture'], fmt='json')
         # badPMTList was inserted by a very clever person into the keras json
         # file. Let's delete it to prevent future keras versions from crashing.
@@ -353,24 +188,14 @@ class PeakPositions1T(strax.Plugin):
             # Nothing to do, and .predict crashes on empty arrays
             return result
 
-        # XENONnT
-        if self.config['nn_file']:
-            _in = peaks['area_per_channel'][peak_mask, 0:self.config['n_top_pmts']]
-            with np.errstate(divide='ignore', invalid='ignore'):
-                _in = _in / np.max(_in, axis=1).reshape(-1, 1)
-            _in = _in.reshape(-1, self.config['n_top_pmts'])
-            _out = self.nn.predict(_in)
+        # Input: normalized hitpatterns in good top PMTs
+        _in = peaks['area_per_channel'][peak_mask, :]
+        _in = _in[:, :self.config['n_top_pmts']][:, self.pmt_mask]
+        with np.errstate(divide='ignore', invalid='ignore'):
+            _in /= _in.sum(axis=1).reshape(-1, 1)
 
-        # XENON1T
-        else:
-            # Input: normalized hitpatterns in good top PMTs
-            _in = peaks['area_per_channel'][peak_mask, :]
-            _in = _in[:, :self.config['n_top_pmts']][:, self.pmt_mask]
-            with np.errstate(divide='ignore', invalid='ignore'):
-                _in /= _in.sum(axis=1).reshape(-1, 1)
-
-            # Output: positions in mm (unfortunately), so convert to cm
-            _out = self.nn.predict(_in) / 10
+        # Output: positions in mm (unfortunately), so convert to cm
+        _out = self.nn.predict(_in) / 10
 
         # Set output in valid rows. Do NOT try result[peak_mask]['x']
         # unless you want all NaN positions (boolean masks make a copy unless
